move widget utilities to a separate module.
